minor function fix

+++ prewitt.m
@@ -7,5 +7,5 @@ function  prewitt_edged  = prewitt(img)
     Jx = conv2(double(img), double(Px), 'same');
     Jy = conv2(double(img), double(Py), 'same');
     %prewitt_edged = Jx + Jy;
-    prewitt_edged = sqrt(Jx.^2 + Jy.^2);
+    prewitt_edged = uint8(sqrt(Jx.^2 + Jy.^2));
 end
